[task 1, stage 5] Вынесла типы для задачи и листа в отдельный файл. Переименовала название для типа листа.
