+++ ReactTasks/task-1/stage-5/src/ts/types.ts
@@ -1,0 +1,13 @@
+export interface Task {
+  id?: string;
+  order?: number;
+  title: string;
+  description: string;
+}
+
+export type ListOfTasks = {
+  id: string;
+  title: string;
+  tasks: Task[];
+};
+
