+++ ReactTasks/task-1/stage-5/src/ts/utils.ts
@@ -1,4 +1,4 @@
-import { Task } from '../data/dataTasks';
+import { Task } from "./types";
 
 export const sortByCardOrder = (a: Task, b: Task) => {
   if (a.order === undefined || b.order === undefined) {
